7.13

+++ mall/models/tools.go
@@ -9,8 +9,8 @@ import (
 )
 
 //时间戳转换成日期
-func UnixToTime(timestamp int) string {
-	t := time.Unix(int64(timestamp), 0)
+func UnixToTime(timestamp int64) string {
+	t := time.Unix(timestamp, 0)
 	return t.Format("2006-01-02 15:04:05")
 }
